* iyudincev/master:
  removed excessive string copying in StringParent::saveToFile
  offsetof
  static analyzer
  some bool results
  changed the types of copy/move conditions to bool
  correction of 93b0b1f
  removed references to LowLevelStr.h
  all SectorSize's are size_t
  refactoring
  correction of 11d6571
  ApplyFileMask: removed limitation of result length, fixed violation of constness
  check for zero bytes allocation in FormatNum

Conflicts:
	src/tools.cpp

+++ src/EngineTools.h
@@ -63,4 +63,4 @@ void setFileTime2(const String & fn, FILETIME * creationTime, FILETIME * lastAcc
 void setFileTime2(HANDLE h, FILETIME * creationTime, FILETIME * lastAccessTime, FILETIME * lastWriteTime);
 size_t Read(HANDLE h, void * buf, size_t size);
 size_t Write(HANDLE h, void * buf, size_t size);
-int GetSectorSize(const String & path);
+size_t GetSectorSize(const String & path);
